// src/pages/dashboard/types.ts
export interface IGridPos {
  h: number;
  w: number;
  x: number;
  y: number;
  i: string;
}

// query interface
export interface ITarget {
  refId: string;
  expr: string; // promQL
  legendFormat: string;
  time?: IRawTimeRange; // 固定时间范围
  step?: number; // 固定时间间隔
  query?: {
    index: string;
    filters: string;
    values: {
      func: string;
    }[];
  };
  legend?: string;
  instant?: boolean;
}

export type IType =
  | 'row'
  | 'timeseries'
  | 'stat'
  | 'table'
  | 'pie'
  | 'hexbin'
  | 'barGauge'
  | 'text'
  | 'gauge'
  | 'iframe';

export interface IValueMapping {
  match: {
    special?: string | number;
    from?: number;
    to?: number;
  };
  result: {
    color: string;
    text: string;
  };
  type: 'range' | 'special' | 'specialValue'; // TODO: 历史原因 special 是固定值，specialValue 是特殊值
}

export interface IThresholds {
  steps: {
    color: string;
    value: number;
    type?: 'base';
  }[];
  // mode: 'absolute' | 'percent'; 目前不支持
  style: 'line'; // 目前只支持 line
}

// 一些通用的配置，不同类型的图表可选择性使用配置
export interface IOptions {
  valueMappings?: IValueMapping[];
  thresholds?: IThresholds;
  xThresholds?: IThresholds;
  standardOptions?: {
    util?: string;
    min?: number;
    max?: number;
    decimals?: number;
    dateFormat?: string;
  };
  legend?: {
    // TODO: 目前不支持这么复杂的自定义
    calcs: string[];
    displayMode: 'list' | 'table' | 'hidden';
    placement: 'right' | 'bottom';
    columns?: string[];
    detailName: string;
    detailUrl: string;
  };
  tooltip?: {
    mode: 'single' | 'all';
    sort: 'none' | 'asc' | 'desc';
  };
}

export interface IOverride {
  matcher: {
    type: 'byName'; // 目前只支持 byName
    value: string;
  };
  properties: {
    [key: string]: any; // standardOptions | valueMappings
  };
}

export interface ILink {
  title: string;
  url: string;
  targetBlank?: boolean;
}
// ...
export interface ITransformation {
  id: 'string';
  options: {
    [key: string]: any;
  };
}

export interface IPanel {
  version: string; // 单个图表面板使用的版本
  id: string;
  name: string;
  links?: ILink[];
  description: string;
  layout: IGridPos;
  datasourceCate?: string; // 5.11.0 新增支持配置数据源类型，默认是 prometheus
  datasourceValue?: number; // 6.x 开始 datasourceName 已经废弃，datasourceValue 即 datasourceId
  targets: ITarget[];
  type: IType;
  options: IOptions;
  custom: any; // 图表
  overrides: IOverride[];
  collapsed?: boolean; // 用于 row 展开收起控制是否显示
  panels?: IPanel[]; // 用于 row 收起时保存子面板
  transformations?: ITransformation[];
  repeat?: string;
  maxPerRow?: number;
  repeatPanelId?: string;
  scopedVars?: any;
}